Add an execute() method to testdriver.
This allows running script using WebDriver in an arbitary browsing
context. That can form the basis for making assertions in cases where
the origin policy prevents such assertions being made purely in
content.

The model is very close to the WebDriver model; there's a sync version
in which one simply returns a value, and an async version in which one
must call a callback which is the last argument to the implied wrapper
function.