Add testdriver messaging API.
This adds send/poll/recv methods to testdriver, which allow sending
messages to other browsing contexts, and polling for messages send to
the current browsing context.

Messages include a data field, which is the message content, and a src
field, which is the uuid of the context in which the message
originated. This allow communicating with cross-origin windows
including those in different browsing context groups.

The transport is a simple server side script, which stores a queue of
messages in the stash.

The implementation is based on dispatcher.js/.py from the COOP
testsuite.