Add reference links to imageBitmapRendering tests and fix flipped orientation test expectation.
Previous CL https://crrev.com/c/3139358 added tests to verify the effect
of the imageOrientation attribute in
ImageBitmapRenderingContext.transferFromImageBitmap. This CL adds links
to those tests' expectations so that they can be run as intended.

This CL additionally fixes one of the above tests which failed because
the test reference was not oriented correctly.

Bug: 1244465
Change-Id: I4508280fb5bc0c60f160c5764f3b61a3f021d135