[@layer] Fix cascade ordering of inline style and layered style
This patch adds a new flag to MatchResult and CascadePriority to
indicate if a declaration is from inline style or other style sheets, so
that we can compare inline style and other layered style correctly.

Without this flag, currently we are treating inline style as regular
declarations in the implicit outer layer, and therefore mistakenly
override it with layered style rules in other style sheets.

Bug: 1095765
Change-Id: I10dbafcf9721fc5b8a5b11edeaeab48c82df86d4