Use the non-hangable-run-end when considering no first-safe position
When executing the ShapeLineBreaker::ShapeLine logic we have a case for
when the first_safe position is bigger than the break_opportunity we
are considering. In that case, we must reshape the whole range.

However, since the change we landed in r873246, we've modified the
BreakIterator's logic to consider breaking opportunities after spaces,
aligning with how ICU is implemented, following the UAX#14 rules. In
order to keep the optimizations to avoid unnecessary reshapes, we also
hold the non-hangable-run-end position, so that we can use it,
depending on the different combinations of values for the CSS Text
properties.

The bug 1248710 is a violation of a DCHECK added to ensure the shaping
algorithm always move forward, to avoid potential infinite-loops.

This CL tries to solve the bug by considering the non-hangable-run-end
position to detected the previously mentioned case of a first-safe
position bigger than the current breaking opportunity's offset.

Bug: 1248710
Change-Id: I5fc2222142d648b74b33033e62270789a84beae3