Use interpolation at the endpoints of animations.
Use interpolation at the endpoints of animations (by removing various
optimizations to return the endpoint values at the endpoints), since
it's needed to get the correct axis for animation of the 'rotate'
property, the correct conversion away from a 'none' value at one
endpoint for the 'rotate', 'scale', and 'translate' properties, and
correct list lengths for list-valued properties that can interpolate
between lists of mismatched lengths, whether by repeating to the least
common multiple length (stroke-dasharray) or filling the shorter list
with no-op or zero values (filter, backdrop-filter, box-shadow,
text-shadow, some registered custom properties).

The changes to translate-composition.html and scale-composition.html
cause Firefox to pass the tests whose expectations are being modified,
and thus, like Chrome, pass the entire file (whereas they cause Safari
to fail additional tests).  The test changes are discussed further in
https://github.com/web-platform-tests/wpt/issues/30377 .

The differences between the test expectations for background-image
(where the test expects discrete animation, per the spec) and for
-webkit-mask-image (where the test expects -webkit-cross-fade()) is
rather suspicious, but I've left the difference as-is for now.

Bug: 1180834
Change-Id: I6320b74b0aff29989a748fab1bff78b91426701e