[LayoutNG] Fieldset overflow clip and fragmentation
If fieldset content exceeded the max block-size due to fragmentation,
we would fail to clip the content correctly (see
overflow-clip-012.html). As a follow-up to CL:2556877 and CL:3141939,
this CL fixes the issue by doing a normal (non-measuring) layout
first, and then relaying out with the max block-size constraint if
the initial result's block size (+ consumed block-size from previous
fragments) exceeded the max.

Bug: 1249489
Change-Id: I8e21f4beb02c631ec677a1967f9867e7fe36d670